refactor(commands): extract SessionControlService; router keeps public forwarders
Move the 6 session-CRUD-with-transport methods (createSessionWithTransport,
removeSessionWithTransport, archiveSessionWithTransport, unarchiveSession,
listNativeSessionsForControl, attachNativeSessionWithTransport) verbatim into
SessionControlService, which composes the TransportInvoker (invoker) and takes
reserveLogicalTransportSession as an injected callback. Router keeps public
forwarders with byte-identical signatures so main.ts/control-service.ts/
console-agent.ts are untouched. Dropped the now-unused resolveRuntimeAgentCommand
import (moved with listNativeSessionsForControl).

Zero behaviour change: 18 golden fixtures byte-identical; mutations S1/S2/S3
reddened only their target scenarios; tsc clean.
