+++ src/commands/command-router.ts
@@ -6,7 +6,6 @@ import { createNoopAppLogger } from "../logging/app-logger";
 import type { SessionService } from "../sessions/session-service";
 import type { AgentCommand, PromptMediaInput, PromptUsage, ReplyQuotaContext, SessionTransport } from "../transport/types";
 import type { AgentSession, ResolvedSession } from "../transport/types";
-import { resolveRuntimeAgentCommand } from "../config/resolve-agent-command";
 import type { PerfSpan } from "../perf/perf-tracer";
 import type { QuotaManager } from "../weixin/messaging/quota-manager.js";
 import { resolveSessionAgentCommandFromIndex, type SessionAgentCommandResolver } from "../transport/acpx-session-index";
@@ -86,6 +85,7 @@ import type {
 } from "./router-types";
 import { renderLaterUnsupportedChannel } from "../scheduled/scheduled-render";
 import { TransportInvoker } from "./transport-invoker";
+import { SessionControlService } from "./session-control-service";
 
 type AutoInstallFn = typeof defaultAutoInstall;
 type DiscoverPathsFn = typeof defaultDiscoverPaths;
@@ -105,6 +105,7 @@ export class CommandRouter {
 
   private readonly activeTurns?: ActiveTurnRegistry;
   private readonly transportInvoker: TransportInvoker;
+  private readonly sessionControl: SessionControlService;
 
   constructor(
     private readonly sessions: SessionService,
@@ -134,6 +135,16 @@ export class CommandRouter {
       resolveSessionAgentCommand: this.resolveSessionAgentCommand,
       autoInstall: (...a) => this.autoInstall(...a),
       discoverPaths: (...a) => this.discoverPaths(...a),
+    });
+    this.sessionControl = new SessionControlService({
+      sessions: this.sessions,
+      transport: this.transport,
+      orchestration: this.orchestration,
+      activeTurns: this.activeTurns,
+      config: this.config,
+      logger: this.logger,
+      invoker: this.transportInvoker,
+      reserveLogicalTransportSession: (ts) => this.reserveLogicalTransportSession(ts),
     });
   }
 
@@ -542,53 +553,7 @@ export class CommandRouter {
     workspace: string,
     model?: string,
   ): Promise<ResolvedSession> {
-    // Refuse to overwrite an existing alias: silently re-pointing it would either
-    // reuse the old transport session (stale history) or orphan it, and a native
-    // session's agent_session_id would be silently dropped. Mirrors handleSessionNew.
-    const existing = this.sessions.getResolvedSessionByInternalAlias(internalAlias);
-    if (existing) {
-      throw new Error(`session "${internalAlias}" already exists`);
-    }
-
-    const session = this.sessions.resolveSession(
-      internalAlias,
-      agent,
-      workspace,
-      `${workspace}:${internalAlias}`,
-    );
-    // An explicit model override must be on the ResolvedSession BEFORE
-    // ensureTransportSession so acpx creates the session under that model
-    // (it carries through as `--model`). Mirrors handleSessionNew.
-    const normalizedModel = model?.trim();
-    if (normalizedModel) {
-      session.model = normalizedModel;
-    }
-    const release = await this.reserveLogicalTransportSession(session.transportSession);
-    try {
-      await this.transportInvoker.ensureTransportSession(session);
-      const exists = await this.transportInvoker.checkTransportSession(session);
-      if (!exists) {
-        throw new Error(`transport session "${session.transportSession}" could not be verified`);
-      }
-      await this.sessions.attachSession(internalAlias, agent, workspace, session.transportSession);
-      if (normalizedModel) {
-        await this.sessions.setSessionModel(internalAlias, normalizedModel);
-      }
-      // Best-effort: a transient refresh failure must not fail a create that has
-      // already succeeded, bound, and verified. Mirrors the chat paths' use of
-      // refreshSessionTransportAgentCommandBestEffort.
-      try {
-        await this.transportInvoker.refreshSessionTransportAgentCommand(internalAlias);
-      } catch (error) {
-        await this.logger.error("session.agent_command_refresh_failed", "failed to refresh session agent command", {
-          alias: internalAlias,
-          error: error instanceof Error ? error.message : String(error),
-        });
-      }
-      return session;
-    } finally {
-      await release();
-    }
+    return this.sessionControl.createSessionWithTransport(internalAlias, agent, workspace, model);
   }
 
   /** Real delete: logical removal + acpx history delete, guarded so a transport
@@ -599,55 +564,7 @@ export class CommandRouter {
     transportTornDown: boolean;
     transportTeardownWarning?: string;
   }> {
-    const session = await this.sessions.getSession(internalAlias);
-    if (!session) {
-      throw new Error(`session "${internalAlias}" does not exist`);
-    }
-    // Both delete entry points (this web/control path and chat `handleSessionRemove`)
-    // MUST enforce the orchestration blocking-task guard + reference purge, or a
-    // coordinator session with in-flight delegated tasks can be irreversibly wiped.
-    if (this.orchestration) {
-      const blocking = await this.orchestration.listSessionBlockingTasks(session.transportSession);
-      if (blocking.length > 0) {
-        throw new Error(`session "${internalAlias}" has ${blocking.length} blocking task(s); cancel them before deleting`);
-      }
-    }
-    const sharedAliasCount = this.sessions.countAliasesSharingTransport(session.transportSession, internalAlias);
-    const { wasActive } = await this.sessions.removeSession(internalAlias);
-
-    if (this.orchestration) {
-      try {
-        await this.orchestration.purgeSessionReferences(session.transportSession);
-      } catch (error) {
-        await this.logger.error("session.orchestration_purge_failed", "failed to purge orchestration references after web remove", {
-          alias: internalAlias,
-          transportSession: session.transportSession,
-          message: error instanceof Error ? error.message : String(error),
-        });
-      }
-    }
-
-    let transportTornDown = false;
-    let transportTeardownWarning: string | undefined;
-    if (sharedAliasCount === 0 && this.transport.deleteSession) {
-      try {
-        await this.transport.deleteSession(session);
-        transportTornDown = true;
-      } catch (error) {
-        transportTeardownWarning = error instanceof Error ? error.message : String(error);
-        await this.logger.error("session.transport_delete_failed", "failed to delete acpx session after logical remove", {
-          alias: internalAlias,
-          transportSession: session.transportSession,
-          message: transportTeardownWarning,
-        });
-      }
-    }
-    return {
-      wasActive,
-      sharedAliasCount,
-      transportTornDown,
-      ...(transportTeardownWarning ? { transportTeardownWarning } : {}),
-    };
+    return this.sessionControl.removeSessionWithTransport(internalAlias);
   }
 
   /** Archive: cancel any in-flight turn and free the warm queue-owner process
@@ -656,51 +573,13 @@ export class CommandRouter {
    *  resumes the same conversation with full history; the first post-archive
    *  prompt cold-starts a fresh queue owner. */
   async archiveSessionWithTransport(internalAlias: string): Promise<void> {
-    const session = await this.sessions.getSession(internalAlias);
-    if (!session) {
-      throw new Error(`session "${internalAlias}" does not exist`);
-    }
-    // Archiving cancels any in-flight turn but deliberately KEEPS the acpx session
-    // alive, so re-prompting later resumes the same conversation with full agent
-    // context + history. Refuse while a turn is in flight so we don't race (and
-    // silently abort) the running prompt.
-    if (this.activeTurns?.isActiveAnywhere(internalAlias)) {
-      throw new Error(`session "${internalAlias}" has a running turn; stop it before archiving`);
-    }
-    const shared = this.sessions.countAliasesSharingTransport(session.transportSession, internalAlias) > 0;
-    if (!shared) {
-      try {
-        await this.transport.cancel(session);
-      } catch {
-        /* best-effort */
-      }
-      // Free the warm queue-owner process now instead of waiting for acpx's TTL to
-      // idle it out. freeWarmProcess kills ONLY the owner process — it does NOT
-      // `sessions close` the record (no `closed` flag), so the session stays open
-      // and the next prompt resumes the same conversation with full history,
-      // repeatably across archive→restore cycles. Best-effort: on failure the
-      // process simply lingers until TTL (the prior behavior), never a regression.
-      try {
-        await this.transport.freeWarmProcess?.(session);
-      } catch (error) {
-        await this.logger.error(
-          "session.free_warm_process_failed",
-          "failed to free warm queue-owner on archive",
-          {
-            alias: internalAlias,
-            transportSession: session.transportSession,
-            message: error instanceof Error ? error.message : String(error),
-          },
-        );
-      }
-    }
-    await this.sessions.setArchived(internalAlias, true);
+    return this.sessionControl.archiveSessionWithTransport(internalAlias);
   }
 
   /** Explicit un-archive (web undo / manual). No process action — it resumes on the
    *  next message via useSession. */
   async unarchiveSession(internalAlias: string): Promise<void> {
-    await this.sessions.setArchived(internalAlias, false);
+    return this.sessionControl.unarchiveSession(internalAlias);
   }
 
   /**
@@ -710,26 +589,7 @@ export class CommandRouter {
    * Returns [] when the transport doesn't support native listing.
    */
   async listNativeSessionsForControl(agent: string, workspace: string): Promise<AgentSession[]> {
-    const listAgentSessions = this.transport.listAgentSessions?.bind(this.transport);
-    if (!listAgentSessions) return [];
-    const agentConfig = this.config?.agents[agent];
-    const workspaceConfig = this.config?.workspaces[workspace];
-    if (!agentConfig || !workspaceConfig) {
-      throw new Error(`unknown agent "${agent}" or workspace "${workspace}"`);
-    }
-    const agentCommand = resolveRuntimeAgentCommand(
-      agentConfig.driver,
-      agentConfig.command,
-      this.config?.transport.preferLocalAgents !== false,
-    );
-    const result = await listAgentSessions({
-      agent,
-      ...(agentCommand ? { agentCommand } : {}),
-      ...(agentConfig.driver ? { driver: agentConfig.driver } : {}),
-      cwd: workspaceConfig.cwd,
-      filterCwd: workspaceConfig.cwd,
-    });
-    return result?.sessions ?? [];
+    return this.sessionControl.listNativeSessionsForControl(agent, workspace);
   }
 
   /**
@@ -745,49 +605,7 @@ export class CommandRouter {
     agentSessionId: string,
     nativeMeta?: { title?: string | null; updatedAt?: string },
   ): Promise<ResolvedSession> {
-    if (!this.transport.resumeAgentSession) {
-      throw new Error("the active transport does not support native sessions");
-    }
-    const existing = this.sessions.getResolvedSessionByInternalAlias(internalAlias);
-    if (existing) {
-      throw new Error(`session "${internalAlias}" already exists`);
-    }
-    const session = this.sessions.resolveSession(
-      internalAlias,
-      agent,
-      workspace,
-      `${workspace}:${internalAlias}`,
-    );
-    const release = await this.reserveLogicalTransportSession(session.transportSession);
-    try {
-      await this.transport.resumeAgentSession(session, agentSessionId);
-      const exists = await this.transportInvoker.checkTransportSession(session);
-      if (!exists) {
-        throw new Error(`transport session "${session.transportSession}" could not be verified`);
-      }
-      await this.sessions.attachNativeSession({
-        alias: internalAlias,
-        agent,
-        workspace,
-        transportSession: session.transportSession,
-        agentSessionId,
-        ...(nativeMeta?.title !== undefined ? { title: nativeMeta.title } : {}),
-        ...(nativeMeta?.updatedAt !== undefined ? { updatedAt: nativeMeta.updatedAt } : {}),
-      });
-      // Best-effort: a transient refresh failure must not fail an attach that already
-      // succeeded, resumed, and verified. Mirrors createSessionWithTransport.
-      try {
-        await this.transportInvoker.refreshSessionTransportAgentCommand(internalAlias);
-      } catch (error) {
-        await this.logger.error("session.native.agent_command_refresh_failed", "failed to refresh native session agent command", {
-          alias: internalAlias,
-          error: error instanceof Error ? error.message : String(error),
-        });
-      }
-      return session;
-    } finally {
-      await release();
-    }
+    return this.sessionControl.attachNativeSessionWithTransport(internalAlias, agent, workspace, agentSessionId, nativeMeta);
   }
 
   private createSessionInteractionOps(perfSpan?: PerfSpan): SessionInteractionOps {
